@@ -1,4 +1,4 @@
-import org.vertx.java.platform.impl.cli.Starter
+import io.vertx.core.Starter
 
 /*
  * Copyright 2012 the original author or authors.
@@ -35,8 +35,8 @@ archivesBaseName = modname
 
 defaultTasks = ['assemble']
 
-sourceCompatibility = '1.7'
-targetCompatibility = '1.7'
+sourceCompatibility = '1.8'
+targetCompatibility = '1.8'
 
 project.ext.moduleName = "$modowner~$modname~$version"
 
@@ -65,26 +65,26 @@ repositories {
 
 dependencies {
     provided "io.vertx:vertx-core:$vertxVersion"
-    provided "io.vertx:vertx-platform:$vertxVersion"
-    
+    provided "io.vertx:vertx-redis-client:$vertxVersion"
+    provided "io.vertx:vertx-web:$vertxVersion"
+
     compile "commons-lang:commons-lang:$commonslangVersion"
     compile "commons-codec:commons-codec:$commonscodecVersion"
-//    compile 'io.vertx:mod-redis-client:1.1.3'
-    
+    compile "org.hamcrest:hamcrest-core:1.3"
+
     runtime "commons-lang:commons-lang:$commonslangVersion"
     runtime "commons-codec:commons-codec:$commonscodecVersion"
-//    runtime 'io.vertx:mod-redis-client:1.1.3'
+    runtime "org.hamcrest:hamcrest-core:1.3"
 
     testCompile "redis.clients:jedis:$jedisVersion"
     testCompile "junit:junit:$junitVersion"
-    testCompile "io.vertx:testtools:$toolsVersion"
+    testCompile "io.vertx:vertx-unit:$vertxVersion"
     testCompile 'com.github.kstyrc:embedded-redis:0.6'
     testCompile 'com.jayway.restassured:rest-assured:2.4.1'
     testCompile 'org.springframework:spring-context:4.1.0.RELEASE'
     testCompile 'org.springframework:spring-jdbc:4.1.0.RELEASE'
-    testCompile 'io.vertx:mod-redis-client:1.1.3'
     testCompile 'com.jayway.awaitility:awaitility-scala:1.6.3'
-    
+
     testRuntime 'com.github.kstyrc:embedded-redis:0.6'
 }
 
@@ -111,7 +111,8 @@ buildscript {
 
   dependencies {
     classpath "io.vertx:vertx-core:$vertxVersion"
-    classpath "io.vertx:vertx-platform:$vertxVersion"
+    classpath "io.vertx:vertx-redis-client:$vertxVersion"
+    classpath "io.vertx:vertx-web:$vertxVersion"
     classpath "io.vertx:vertx-hazelcast:$vertxVersion"
 
     // Extra classpath for Vert.x
