@@ -1,24 +1,25 @@
 package li.chee.vertx.reststorage;
 
-import static com.jayway.awaitility.Duration.TWO_SECONDS;
-import com.jayway.restassured.RestAssured;
-import org.junit.Before;
+import io.vertx.ext.unit.Async;
+import io.vertx.ext.unit.TestContext;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 
+import static com.jayway.awaitility.Awaitility.await;
+import static com.jayway.awaitility.Duration.TWO_SECONDS;
 import static com.jayway.restassured.RestAssured.*;
-import static com.jayway.restassured.RestAssured.given;
-import static com.jayway.awaitility.Awaitility.*;
 import static org.hamcrest.core.IsEqual.equalTo;
-import static org.vertx.testtools.VertxAssert.testComplete;
 
+@RunWith(VertxUnitRunner.class)
 public class ExpirationTest extends AbstractTestCase {
 
     @Test
-    public void testPutInvalidExpireFloat() {
-
+    public void testPutInvalidExpireFloat(TestContext context) {
+        Async async = context.async();
         given().
                 header("x-expire-after", "1.22").
                 body("{ \"foo\": \"bar1\" }").
@@ -28,12 +29,12 @@ public class ExpirationTest extends AbstractTestCase {
                 assertThat().statusCode(400);
 
         when().get("expireaftertwoseconds").then().assertThat().statusCode(404);
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPutInvalidExpireNaN() {
-
+    public void testPutInvalidExpireNaN(TestContext context) {
+        Async async = context.async();
         given().
                 header("x-expire-after", "asdfasdf").
                 body("{ \"foo\": \"bar1\" }").
@@ -43,12 +44,12 @@ public class ExpirationTest extends AbstractTestCase {
                 assertThat().statusCode(400);
 
         when().get("invalidExpireNan").then().assertThat().statusCode(404);
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPutExpireAfterOneSecond() {
-
+    public void testPutExpireAfterOneSecond(TestContext context) {
+        Async async = context.async();
         given().
                 header("x-expire-after", "1").
                 body("{ \"foo\": \"bar1\" }").
@@ -62,12 +63,12 @@ public class ExpirationTest extends AbstractTestCase {
                 return get("expireaftertwoseconds").statusCode();
             }
         }, equalTo(404));
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPutAfterExpiration() {
-
+    public void testPutAfterExpiration(TestContext context) {
+        Async async = context.async();
         given().
                 header("x-expire-after", "1").
                 body("{ \"foo\": \"bar1\" }").
@@ -92,12 +93,12 @@ public class ExpirationTest extends AbstractTestCase {
                 assertThat().statusCode(200);
 
         when().get("putafterexpiration").then().statusCode(200).body("foo", equalTo("bar2"));
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPutAfterImmediateExpiration() {
-
+    public void testPutAfterImmediateExpiration(TestContext context) {
+        Async async = context.async();
         given().
                 header("x-expire-after", "0").
                 body("{ \"foo\": \"bar1\" }").
@@ -107,11 +108,7 @@ public class ExpirationTest extends AbstractTestCase {
                 assertThat().statusCode(200);
 
 
-        await().atMost(TWO_SECONDS).until(new Callable<Integer>() {
-            public Integer call() throws Exception {
-                return get("expireimmediatly").statusCode();
-            }
-        }, equalTo(404));
+        await().atMost(TWO_SECONDS).until(() -> get("expireimmediatly").statusCode(), equalTo(404));
 
         given().
                 header("x-expire-after", "10").
@@ -122,12 +119,12 @@ public class ExpirationTest extends AbstractTestCase {
                 assertThat().statusCode(200);
 
         when().get("/expireimmediatly").then().statusCode(200).body("foo", equalTo("bar2"));
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPutMultipleBranchesAfterImmediateExpiration() {
-
+    public void testPutMultipleBranchesAfterImmediateExpiration(TestContext context) {
+        Async async = context.async();
         given().
                 header("x-expire-after", "0").
                 body("{ \"foo\": \"bar1\" }").
@@ -153,11 +150,7 @@ public class ExpirationTest extends AbstractTestCase {
                 assertThat().statusCode(200);
 
 
-        await().atMost(TWO_SECONDS).until(new Callable<Integer>() {
-            public Integer call() throws Exception {
-                return get("resexpireimmediatly/branch1").statusCode();
-            }
-        }, equalTo(404));
+        await().atMost(TWO_SECONDS).until(() -> get("resexpireimmediatly/branch1").statusCode(), equalTo(404));
 
         when().get("resexpireimmediatly/branch2").then().assertThat().statusCode(404);
         when().get("resexpireimmediatly/branch3").then().assertThat().statusCode(404);
@@ -188,12 +181,12 @@ public class ExpirationTest extends AbstractTestCase {
         when().get("resexpireimmediatly/branch1").then().statusCode(200).body("foo", equalTo("bar11"));
         when().get("resexpireimmediatly/branch2").then().statusCode(200).body("foo", equalTo("bar22"));
         when().get("resexpireimmediatly/branch3").then().statusCode(200).body("foo", equalTo("bar33"));
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPutWithNoExpiryAfterImmediateExpiration() {
-
+    public void testPutWithNoExpiryAfterImmediateExpiration(TestContext context) {
+        Async async = context.async();
         given().
                 header("x-expire-after", "0").
                 body("{ \"foo\": \"bar1\" }").
@@ -212,6 +205,6 @@ public class ExpirationTest extends AbstractTestCase {
                 assertThat().statusCode(200);
 
         when().get("resexpireimmediatly/branch1").then().statusCode(200).body("foo", equalTo("bar11"));
-        testComplete();
+        async.complete();
     }
 }
