@@ -1,19 +1,17 @@
 package li.chee.vertx.reststorage;
 
 import io.netty.handler.codec.http.QueryStringDecoder;
-import org.vertx.java.core.*;
-import org.vertx.java.core.buffer.Buffer;
-import org.vertx.java.core.eventbus.Message;
-import org.vertx.java.core.http.*;
-import org.vertx.java.core.json.JsonArray;
-import org.vertx.java.core.json.JsonObject;
-import org.vertx.java.core.net.NetSocket;
+import io.vertx.core.*;
+import io.vertx.core.buffer.Buffer;
+import io.vertx.core.eventbus.Message;
+import io.vertx.core.http.*;
+import io.vertx.core.json.JsonArray;
+import io.vertx.core.json.JsonObject;
+import io.vertx.core.net.NetSocket;
+import io.vertx.core.net.SocketAddress;
 
 import javax.net.ssl.SSLPeerUnverifiedException;
 import javax.security.cert.X509Certificate;
-import java.net.InetSocketAddress;
-import java.net.URI;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -25,7 +23,7 @@ import java.util.Map;
 public class EventBusAdapter {
 
     public void init(final Vertx vertx, String address, final Handler<HttpServerRequest> requestHandler) {
-        vertx.eventBus().registerHandler(address, new Handler<Message<Buffer>>() {
+        vertx.eventBus().consumer(address, new Handler<Message<Buffer>>() {
             @Override
             public void handle(Message<Buffer> message) {
                 requestHandler.handle(new MappedHttpServerRequest(vertx, message));
@@ -36,7 +34,7 @@ public class EventBusAdapter {
     private class MappedHttpServerRequest implements HttpServerRequest {
         private Vertx vertx;
         private Buffer requestPayload;
-        private String method;
+        private HttpMethod method;
         private String uri;
         private String path;
         private String query;
@@ -53,16 +51,24 @@ public class EventBusAdapter {
             Buffer buffer = message.body();
             int headerLength =  buffer.getInt(0);
             JsonObject header = new JsonObject(buffer.getString(4,headerLength+4));
-            method = header.getString("method");
+            method = httpMethodFromHeader(header);
             uri = header.getString("uri");
             requestPayload = buffer.getBuffer(headerLength+4, buffer.length());
 
-            JsonArray headerArray = header.getArray("headers");
+            JsonArray headerArray = header.getJsonArray("headers");
             if(headerArray != null) {
                 requestHeaders = fromJson(headerArray);
             } else {
-                requestHeaders = new CaseInsensitiveMultiMap();
+                requestHeaders = new CaseInsensitiveHeaders();
             }
+        }
+
+        private HttpMethod httpMethodFromHeader(JsonObject header){
+            String method = header.getString("method");
+            if(method != null){
+                return HttpMethod.valueOf(method.toUpperCase());
+            }
+            return null;
         }
 
         @Override
@@ -71,7 +77,7 @@ public class EventBusAdapter {
         }
 
         @Override
-        public String method() {
+        public HttpMethod method() {
             return method;
         }
 
@@ -103,8 +109,8 @@ public class EventBusAdapter {
 
                     private int statusCode;
                     private String statusMessage;
-                    private MultiMap responseHeaders = new CaseInsensitiveMultiMap();
-                    private Buffer responsePayload = new Buffer();
+                    private MultiMap responseHeaders = new CaseInsensitiveHeaders();
+                    private Buffer responsePayload = Buffer.buffer();
 
                     @Override
                     public int getStatusCode() {
@@ -209,19 +215,24 @@ public class EventBusAdapter {
 
                     @Override
                     public HttpServerResponse write(String s, String s2) {
-                        responsePayload.appendBuffer(new Buffer(s, s2));
+                        responsePayload.appendBuffer(Buffer.buffer(s, s2));
                         return this;
                     }
 
                     @Override
                     public HttpServerResponse write(String s) {
-                        responsePayload.appendBuffer(new Buffer(s));
+                        responsePayload.appendBuffer(Buffer.buffer(s));
                         return this;
                     }
 
                     @Override
+                    public HttpServerResponse writeContinue() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    @Override
                     public void end(String s) {
-                        write(new Buffer(s));
+                        write(Buffer.buffer(s));
                         end();
                     }
 
@@ -244,11 +255,11 @@ public class EventBusAdapter {
                             statusCode = 200;
                             statusMessage = "OK";
                         }
-                        header.putNumber("statusCode", statusCode);
-                        header.putString("statusMessage", statusMessage);
-                        header.putArray("headers", toJson(responseHeaders));
-                        Buffer bufferHeader = new Buffer(header.encode());
-                        Buffer response = new Buffer(4+bufferHeader.length()+responsePayload.length());
+                        header.put("statusCode", statusCode);
+                        header.put("statusMessage", statusMessage);
+                        header.put("headers", toJson(responseHeaders));
+                        Buffer bufferHeader = Buffer.buffer(header.encode());
+                        Buffer response = Buffer.buffer(4+bufferHeader.length()+responsePayload.length());
                         response.setInt(0, bufferHeader.length()).appendBuffer(bufferHeader).appendBuffer(responsePayload);
                         message.reply(response);
                     }
@@ -259,9 +270,10 @@ public class EventBusAdapter {
                     }
 
                     @Override
-                    public HttpServerResponse sendFile(String s, String s2) {
-                        throw new UnsupportedOperationException();
-                    }
+                    public HttpServerResponse sendFile(String filename, long offset) { throw new UnsupportedOperationException(); }
+
+                    @Override
+                    public HttpServerResponse sendFile(String filename, long offset, long length) { throw new UnsupportedOperationException(); }
 
                     @Override
                     public HttpServerResponse sendFile(String s, Handler<AsyncResult<Void>> asyncResultHandler) {
@@ -269,18 +281,51 @@ public class EventBusAdapter {
                     }
 
                     @Override
-                    public HttpServerResponse sendFile(String s, String s2, Handler<AsyncResult<Void>> asyncResultHandler) {
+                    public HttpServerResponse sendFile(String filename, long offset, Handler<AsyncResult<Void>> resultHandler) {
                         throw new UnsupportedOperationException();
                     }
 
                     @Override
-                    public void close() {
+                    public HttpServerResponse sendFile(String filename, long offset, long length, Handler<AsyncResult<Void>> resultHandler) {
+                        throw new UnsupportedOperationException();
+                    }
 
+                    @Override
+                    public void close() {}
+
+                    @Override
+                    public boolean ended() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    @Override
+                    public boolean closed() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    @Override
+                    public boolean headWritten() {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    @Override
+                    public HttpServerResponse headersEndHandler(Handler<Void> handler) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    @Override
+                    public HttpServerResponse bodyEndHandler(Handler<Void> handler) {
+                        throw new UnsupportedOperationException();
+                    }
+
+                    @Override
+                    public long bytesWritten() {
+                        throw new UnsupportedOperationException();
                     }
 
                     @Override
                     public HttpServerResponse setWriteQueueMaxSize(int i) {
-                        return this;
+                        throw new UnsupportedOperationException();
                     }
 
                     @Override
@@ -308,11 +353,21 @@ public class EventBusAdapter {
         }
 
         @Override
+        public String getHeader(String headerName) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public String getHeader(CharSequence headerName) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
         public MultiMap params() {
             if (params == null) {
                 QueryStringDecoder queryStringDecoder = new QueryStringDecoder(uri());
                 Map<String, List<String>> prms = queryStringDecoder.parameters();
-                params = new CaseInsensitiveMultiMap();
+                params = new CaseInsensitiveHeaders();
                 if (!prms.isEmpty()) {
                     for (Map.Entry<String, List<String>> entry: prms.entrySet()) {
                         params.add(entry.getKey(), entry.getValue());
@@ -323,12 +378,17 @@ public class EventBusAdapter {
         }
 
         @Override
-        public InetSocketAddress remoteAddress() {
+        public String getParam(String paramName) {
             throw new UnsupportedOperationException();
         }
 
         @Override
-        public InetSocketAddress localAddress() {
+        public SocketAddress remoteAddress() {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public SocketAddress localAddress() {
             throw new UnsupportedOperationException();
         }
 
@@ -338,18 +398,14 @@ public class EventBusAdapter {
         }
 
         @Override
-        public URI absoluteURI() {
+        public String absoluteURI() {
             return null;
         }
 
         @Override
         public HttpServerRequest bodyHandler(final Handler<Buffer> bodyHandler) {
-            final Buffer body = new Buffer();
-            dataHandler(new Handler<Buffer>() {
-                public void handle(Buffer buff) {
-                    body.appendBuffer(buff);
-                }
-            });
+            final Buffer body = Buffer.buffer();
+            handler(body::appendBuffer);
             endHandler(new VoidHandler() {
                 public void handle() {
                     bodyHandler.handle(body);
@@ -364,8 +420,13 @@ public class EventBusAdapter {
         }
 
         @Override
-        public HttpServerRequest expectMultiPart(boolean b) {
+        public HttpServerRequest setExpectMultipart(boolean expect) {
             return this;
+        }
+
+        @Override
+        public boolean isExpectMultipart() {
+            throw new UnsupportedOperationException();
         }
 
         @Override
@@ -379,25 +440,25 @@ public class EventBusAdapter {
         }
 
         @Override
+        public String getFormAttribute(String attributeName) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public ServerWebSocket upgrade() {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public boolean isEnded() {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
         public HttpServerRequest endHandler(Handler<Void> voidHandler) {
             endHandler = voidHandler;
             if(requestPayload == null) {
                 endHandler.handle(null);
-            }
-            return this;
-        }
-
-        @Override
-        public HttpServerRequest dataHandler(Handler<Buffer> bufferHandler) {
-            if(requestPayload != null) {
-                dataHandler = bufferHandler;
-                vertx.runOnContext(new Handler<Void>() {
-                    @Override
-                    public void handle(Void aVoid) {
-                        dataHandler.handle(requestPayload);
-                        endHandler.handle(null);
-                    }
-                });
             }
             return this;
         }
@@ -416,25 +477,35 @@ public class EventBusAdapter {
         public HttpServerRequest exceptionHandler(Handler<Throwable> throwableHandler) {
             return this;
         }
+
+        @Override
+        public HttpServerRequest handler(Handler<Buffer> bufferHandler) {
+            if(requestPayload != null) {
+                dataHandler = bufferHandler;
+                vertx.runOnContext(aVoid -> {
+                    dataHandler.handle(requestPayload);
+                    endHandler.handle(null);
+                });
+            }
+            return this;
+        }
     }
 
     public static JsonArray toJson(MultiMap multiMap) {
         JsonArray result = new JsonArray();
         for(Map.Entry<String, String> entry: multiMap.entries()) {
-            result.addArray(new JsonArray().add(entry.getKey()).add(entry.getValue()));
+            result.add(new JsonArray().add(entry.getKey()).add(entry.getValue()));
         }
         return result;
     }
 
     public static MultiMap fromJson(JsonArray json) {
-        MultiMap result = new CaseInsensitiveMultiMap();
-        Iterator<Object> it = json.iterator();
-        while(it.hasNext()) {
-            Object next = it.next();
-            if(next instanceof JsonArray) {
-                JsonArray pair = (JsonArray)next;
-                if(pair.size() == 2) {
-                    result.add(pair.get(0).toString(), pair.get(1).toString());
+        MultiMap result = new CaseInsensitiveHeaders();
+        for (Object next : json) {
+            if (next instanceof JsonArray) {
+                JsonArray pair = (JsonArray) next;
+                if (pair.size() == 2) {
+                    result.add(pair.getString(0), pair.getString(1));
                 }
             }
         }
