@@ -1,36 +1,41 @@
 package li.chee.vertx.reststorage;
 
 import com.jayway.restassured.RestAssured;
-import org.junit.After;
-import org.junit.Before;
+import io.vertx.ext.unit.Async;
+import io.vertx.ext.unit.TestContext;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 import static com.jayway.restassured.RestAssured.*;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.hasItems;
-import static org.vertx.testtools.VertxAssert.testComplete;
 
+@RunWith(VertxUnitRunner.class)
 public class CrudTest extends AbstractTestCase {
 
     @Test
-    public void testPutGetDelete() {
+    public void testPutGetDelete(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar\" }").put("res");
         when().get("res").then().assertThat().body("foo", equalTo("bar"));
         delete("res");
         when().get("res").then().assertThat().statusCode(404);
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testList() {
+    public void testList(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar\" }").put("resources/res1");
         with().body("{ \"foo\": \"bar2\" }").put("resources/res2");
         when().get("resources").then().assertThat().body("resources", hasItems("res1", "res2"));
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testMerge() {
+    public void testMerge(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar\", \"hello\": \"world\" }").put("res");
         with().
                 param("merge", "true").
@@ -40,12 +45,12 @@ public class CrudTest extends AbstractTestCase {
                 statusCode(200).
                 body("foo", equalTo("bar2")).
                 body("hello", equalTo("world"));
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testTwoBranchesDeleteOnLeafOfOneBranch() {
-
+    public void testTwoBranchesDeleteOnLeafOfOneBranch(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar1\" }").put("branch1/res/leaf");
         with().body("{ \"foo\": \"bar2\" }").put("branch2/res/leaf");
 
@@ -61,12 +66,12 @@ public class CrudTest extends AbstractTestCase {
         when().get("branch2/res/leaf").then().assertThat().statusCode(404);
 
         delete("branch2/res/leaf");
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testTwoBranchesDeleteOnNodeUnderLeafOfOneBranch() {
-
+    public void testTwoBranchesDeleteOnNodeUnderLeafOfOneBranch(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar1\" }").put("branch1/res/res1/leaf");
         with().body("{ \"foo\": \"bar2\" }").put("branch2/res/res2/leaf");
 
@@ -82,12 +87,12 @@ public class CrudTest extends AbstractTestCase {
         when().get("branch2/res").then().assertThat().statusCode(404);
         when().get("branch2/res/res2").then().assertThat().statusCode(404);
         when().get("branch2/res/res2/leaf").then().assertThat().statusCode(404);
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testTwoBranchesDeleteOnNodeAfterBranch() {
-
+    public void testTwoBranchesDeleteOnNodeAfterBranch(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar1\" }").put("branch1/res/res1/leaf");
         with().body("{ \"foo\": \"bar2\" }").put("branch2/res/res2/leaf");
 
@@ -102,12 +107,12 @@ public class CrudTest extends AbstractTestCase {
         when().get("branch1/res").then().assertThat().statusCode(404);
         when().get("branch1/res/leaf").then().assertThat().statusCode(404);
         when().get("branch1/res/res1/leaf").then().assertThat().statusCode(404);
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testTwoBranchesDeleteOnRoot() {
-
+    public void testTwoBranchesDeleteOnRoot(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar\" }").put("node/node/node/branch1/res/res1/leaf");
         with().body("{ \"foo\": \"bar\" }").put("node/node/node/branch2/res/res2/leaf");
 
@@ -125,13 +130,13 @@ public class CrudTest extends AbstractTestCase {
         when().get("node/node/node/branch1/res").then().assertThat().statusCode(404);
         when().get("node/node/node/branch1/res/leaf").then().assertThat().statusCode(404);
         when().get("node/node/node/branch1/res/res1/leaf").then().assertThat().statusCode(404);
-        testComplete();
+        async.complete();
     }
 
 
     @Test
-    public void testThreeBranchesDeleteOnLeafOfOneBranch() {
-
+    public void testThreeBranchesDeleteOnLeafOfOneBranch(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar1\" }").put("node/branch1/res/leaf");
         with().body("{ \"foo\": \"bar2\" }").put("node/branch2/res/leaf");
         with().body("{ \"foo\": \"bar3\" }").put("node/branch3/res/leaf");
@@ -149,12 +154,12 @@ public class CrudTest extends AbstractTestCase {
         when().get("node/branch3").then().assertThat().statusCode(404);
         when().get("node/branch3/res").then().assertThat().statusCode(404);
         when().get("node/branch3/res/leaf").then().assertThat().statusCode(404);
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testThreeBranchesDeleteOnNodeUnderLeafOfOneBranch() {
-
+    public void testThreeBranchesDeleteOnNodeUnderLeafOfOneBranch(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar1\" }").put("branch1/res/res1/leaf");
         with().body("{ \"foo\": \"bar2\" }").put("branch2/res/res2/leaf");
         with().body("{ \"foo\": \"bar3\" }").put("branch3/res/res3/leaf");
@@ -175,12 +180,12 @@ public class CrudTest extends AbstractTestCase {
         when().get("/branch3/res").then().assertThat().statusCode(200);
         when().get("/branch3/res/res3").then().assertThat().statusCode(200);
         when().get("/branch3/res/res3/leaf").then().assertThat().statusCode(200).body("foo", equalTo("bar3"));
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testThreeBranchesDeleteOnNodeAfterBranch() {
-
+    public void testThreeBranchesDeleteOnNodeAfterBranch(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar1\" }").put("branch1/res/res1/leaf");
         with().body("{ \"foo\": \"bar2\" }").put("branch2/res/res2/leaf");
         with().body("{ \"foo\": \"bar3\" }").put("branch3/res/res3/leaf");
@@ -202,12 +207,12 @@ public class CrudTest extends AbstractTestCase {
         when().get("branch3/res").then().assertThat().statusCode(200);
         when().get("branch3/res/res3").then().assertThat().statusCode(200);
         when().get("branch3/res/res3/leaf").then().assertThat().statusCode(200).body("foo", equalTo("bar3"));
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testThreeBranchesDeleteOnRoot() {
-
+    public void testThreeBranchesDeleteOnRoot(TestContext context) {
+        Async async = context.async();
         with().body("{ \"foo\": \"bar\" }").put("node/node/node/branch1/res/res1/leaf");
         with().body("{ \"foo\": \"bar\" }").put("node/node/node/branch2/res/res2/leaf");
         with().body("{ \"foo\": \"bar\" }").put("node/node/node/branch3/res/res3/leaf");
@@ -230,6 +235,6 @@ public class CrudTest extends AbstractTestCase {
         when().get("node/node/node/branch3/res").then().assertThat().statusCode(404);
         when().get("node/node/node/branch3/res/leaf").then().assertThat().statusCode(404);
         when().get("node/node/node/branch3/res/res3/leaf").then().assertThat().statusCode(404);
-        testComplete();
+        async.complete();
     }
 }
