@@ -1,119 +1,115 @@
 package li.chee.vertx.reststorage;
 
-import static org.vertx.testtools.VertxAssert.testComplete;
-
+import io.vertx.core.DeploymentOptions;
+import io.vertx.core.Handler;
+import io.vertx.core.Vertx;
+import io.vertx.core.http.HttpClient;
+import io.vertx.core.http.HttpClientOptions;
+import io.vertx.core.http.HttpClientRequest;
+import io.vertx.core.http.HttpClientResponse;
+import io.vertx.core.json.JsonObject;
+import io.vertx.core.logging.Logger;
+import io.vertx.core.logging.LoggerFactory;
+import io.vertx.ext.unit.Async;
+import io.vertx.ext.unit.TestContext;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
-import org.vertx.java.core.AsyncResult;
-import org.vertx.java.core.AsyncResultHandler;
-import org.vertx.java.core.Handler;
-import org.vertx.java.core.buffer.Buffer;
-import org.vertx.java.core.http.HttpClient;
-import org.vertx.java.core.http.HttpClientRequest;
-import org.vertx.java.core.http.HttpClientResponse;
-import org.vertx.java.core.json.JsonObject;
-import org.vertx.java.core.logging.Logger;
-import org.vertx.testtools.TestVerticle;
+import org.junit.runner.RunWith;
 
-public class IntegrationTest extends TestVerticle {
+@RunWith(VertxUnitRunner.class)
+public class IntegrationTest {
 
-    private Logger log;
+    Vertx vertx;
+    HttpClient client;
+    private Logger log = LoggerFactory.getLogger(IntegrationTest.class);
 
-    @Override
-    public void start() {
+    @Before
+    public void setUp(TestContext context) {
+        vertx = Vertx.vertx();
 
-        log = container.logger();
+        JsonObject storageConfig = new JsonObject();
+        storageConfig.put("storage", "redis");
+        storageConfig.put("storageAddress", "we-love-to-put");
+        storageConfig.put("prefix", "/test");
+        vertx.deployVerticle("li.chee.vertx.reststorage.RestStorageMod", new DeploymentOptions().setConfig(storageConfig), context.asyncAssertSuccess());
+    }
 
-        // Make sure we call initialize() - this sets up the assert stuff so assert
-        // functionality works correctly
-        initialize();
-        // Deploy the module - the System property `vertx.modulename` will contain
-        // the name of the module so you
-        // don't have to hardecode it in your tests
-        final String moduleName = System.getProperty("vertx.modulename");
-        JsonObject config = new JsonObject().putString("storageAddress", "we-love-to-put").putString("prefix", "/test").putString("root", "dogs");
-        container.deployModule(moduleName, config, 1, new AsyncResultHandler<String>() {
-            public void handle(AsyncResult<String> event) {
-                log.info("success of deployment of  module " + moduleName + ": " + event.result());
-               	startTests();
-            }
-        });
+    @After
+    public void tearDown(TestContext context) {
+        vertx.close(context.asyncAssertSuccess());
+    }
+
+    @Test
+    public void testSimple(TestContext context) {
+        Async async = context.async();
+        client = vertx.createHttpClient(new HttpClientOptions().setDefaultPort(8989));
+        HttpClientRequest request = client.get("/test/", new PrintHandler(step2(async), async));
+        request.headers().add("Accept", "text/html");
+        request.end();
     }
 
     class PrintHandler implements Handler<HttpClientResponse> {
         boolean complete;
         Handler<Void> next;
+        Async async;
         
         public PrintHandler() {            
         }
         
-        public PrintHandler(boolean complete) {
+        public PrintHandler(boolean complete, Async async) {
             this.complete = complete;
+            this.async = async;
         }
         
-        public PrintHandler(Handler<Void> next) {
+        public PrintHandler(Handler<Void> next, Async async) {
             this.next = next;
+            this.async = async;
         }
 
         public void handle(HttpClientResponse response) {
-            System.out.println(response.statusCode());
-            response.bodyHandler(new Handler<Buffer>() {
-                public void handle(Buffer body) {
-                    System.out.println(body.toString());
-                    if(next != null) {
-                        next.handle(null);
-                    }
-                    if (complete) {
-                        testComplete();
-                    }
+            log.info(response.statusCode());
+            response.bodyHandler(body -> {
+                log.info(body.toString());
+                if(next != null) {
+                    next.handle(null);
+                }
+                if (complete) {
+                    async.complete();
                 }
             });
         }
     }
 
-    HttpClient client;
-    
-    @Test
-    public void testSimple() {
-        client = vertx.createHttpClient().setPort(8989);
-        HttpClientRequest request = client.get("/test/", new PrintHandler(step2()));
-        request.headers().add("Accept", "text/html");
-        request.end();
-    }
-    
-    private Handler<Void> step2() {
-        return new Handler<Void>() {
-            public void handle(Void event) {
-                StringBuilder content = new StringBuilder("{");
-                for(int i=0; i<10000; i++) {
-                     content.append(" \"hello\": \"world\",\n");
-                }
-                content.append("\"hello\": \"world\" }");
-                HttpClientRequest request = client.put("/test/hello", new PrintHandler(step3()));
-                request.headers().add("Content-Length", ""+content.length());
-                request.write(content.toString());
-                request.end();
-            }            
+    private Handler<Void> step2(Async async) {
+        return event -> {
+            StringBuilder content = new StringBuilder("{");
+            for(int i=0; i<10000; i++) {
+                 content.append(" \"hello\": \"world\",\n");
+            }
+            content.append("\"hello\": \"world\" }");
+            HttpClientRequest request = client.put("/test/hello", new PrintHandler(step3(async), async));
+            request.headers().add("Content-Length", ""+content.length());
+            request.write(content.toString());
+            request.end();
         };
     }
     
-    private Handler<Void> step3() {
-        return new Handler<Void>() {
-            public void handle(Void event) {
-                String content = "{ \"hello\": \"world\" }";
-                HttpClientRequest request = client.put("/test/world", new PrintHandler(step4()));
-                request.headers().add("Content-Length", ""+content.length());
-                request.write(content);
-                request.end();
-            }            
+    private Handler<Void> step3(Async async) {
+        return event -> {
+            String content = "{ \"hello\": \"world\" }";
+            HttpClientRequest request = client.put("/test/world", new PrintHandler(step4(async), async));
+            request.headers().add("Content-Length", ""+content.length());
+            request.write(content);
+            request.end();
         };
     }
     
-    private Handler<Void> step4() {
-        return new Handler<Void>() {
-            public void handle(Void event) {
-                HttpClientRequest request = client.delete("/test/world", new PrintHandler(true));
-                request.end();
-            }            
+    private Handler<Void> step4(Async async) {
+        return event -> {
+            HttpClientRequest request = client.delete("/test/world", new PrintHandler(true, async));
+            request.end();
         };
     }
 }
