@@ -1,18 +1,20 @@
 package li.chee.vertx.reststorage;
 
 import com.jayway.restassured.RestAssured;
-import org.junit.Before;
+import io.vertx.ext.unit.Async;
+import io.vertx.ext.unit.TestContext;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 import static com.jayway.restassured.RestAssured.*;
-import static org.vertx.testtools.VertxAssert.assertEquals;
-import static org.vertx.testtools.VertxAssert.testComplete;
 
+@RunWith(VertxUnitRunner.class)
 public class PathLevelTest extends AbstractTestCase {
 
     @Test
-    public void testTryToPutResourceOverCollection() {
-
+    public void testTryToPutResourceOverCollection(TestContext context) {
+        Async async = context.async();
         RestAssured.basePath = "";
         with().put("/tests/crush/test1/test2/test3");
         // here we assume, that on the path server is already a collection
@@ -22,11 +24,12 @@ public class PathLevelTest extends AbstractTestCase {
                 put("/tests").
                 then().
                 assertThat().statusCode(405);
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPut4levels() {
+    public void testPut4levels(TestContext context) {
+        Async async = context.async();
         // cleanup
         delete("");
 
@@ -38,29 +41,29 @@ public class PathLevelTest extends AbstractTestCase {
                 assertThat().statusCode(200);
 
         // test level 1 with and without trailing slash
-        assertEquals("[test2/]", get("/tests/crush/test1").body().jsonPath().get("test1").toString());
-        assertEquals("[test2/]", get("/tests/crush/test1").body().jsonPath().get("test1").toString());
-        assertEquals("[test2/]", get("/tests/crush/test1/").body().jsonPath().get("test1").toString());
+        context.assertEquals("[test2/]", get("/tests/crush/test1").body().jsonPath().get("test1").toString());
+        context.assertEquals("[test2/]", get("/tests/crush/test1").body().jsonPath().get("test1").toString());
+        context.assertEquals("[test2/]", get("/tests/crush/test1/").body().jsonPath().get("test1").toString());
 
         // test level 2 with and without trailing slash
-        assertEquals("[test3/]", get("/tests/crush/test1/test2").body().jsonPath().get("test2").toString());
-        assertEquals("[test3/]", get("/tests/crush/test1/test2/").body().jsonPath().get("test2").toString());
+        context.assertEquals("[test3/]", get("/tests/crush/test1/test2").body().jsonPath().get("test2").toString());
+        context.assertEquals("[test3/]", get("/tests/crush/test1/test2/").body().jsonPath().get("test2").toString());
 
         // test level 3 with and without trailing slash
-        assertEquals("[test4]", get("/tests/crush/test1/test2/test3").body().jsonPath().get("test3").toString());
-        assertEquals("[test4]", get("/tests/crush/test1/test2/test3/").body().jsonPath().get("test3").toString());
+        context.assertEquals("[test4]", get("/tests/crush/test1/test2/test3").body().jsonPath().get("test3").toString());
+        context.assertEquals("[test4]", get("/tests/crush/test1/test2/test3/").body().jsonPath().get("test3").toString());
 
         // test4 level
-        assertEquals("{ \"foo\": \"bar\" }", get("/tests/crush/test1/test2/test3/test4").body().asString());
+        context.assertEquals("{ \"foo\": \"bar\" }", get("/tests/crush/test1/test2/test3/test4").body().asString());
 
         // cleanup
         delete("");
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPutResourceOverCollection() {
-
+    public void testPutResourceOverCollection(TestContext context) {
+        Async async = context.async();
         given().
                 body("{ \"foo\": \"bar\" }").
                 when().
@@ -77,12 +80,12 @@ public class PathLevelTest extends AbstractTestCase {
 
         // cleanup
         delete("");
-        testComplete();
+        async.complete();
     }
 
     @Test
-    public void testPutCollectionOverResource() {
-
+    public void testPutCollectionOverResource(TestContext context) {
+        Async async = context.async();
         given().
                 body("{ \"foo\": \"bar\" }").
                 when().
@@ -99,6 +102,6 @@ public class PathLevelTest extends AbstractTestCase {
 
         // cleanup
         delete("");
-        testComplete();
+        async.complete();
     }
 }
